requirements.txt
streamlit==1.35.0
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm==4.66.4
protobuf<4.21.3
pandas>=1.1.4
seaborn>=0.11.0
tensorboard>=2.4.1
ipython
psutil
thop